Officially moved to bot.commands
Officially moved the bot to bot.commands for ease of use with Discord RBAC to be able to restrict commands to certain roles. Before we had to put the bot in a secret channel. Now the Guild does not need to hide the bot at all and instead adds the role(s) that are needed to run the command.

Also removed deprecated code but kept old code for future reference.